ref(issues): Share native frame variable types
Move the native variable model alongside Frame so renderers and future extraction adapters can share it.
